api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
towhee==1.1.0
pillow==10.0.1
numpy==1.24.3
pymilvus==2.4.0
supabase==1.0.3
python-multipart==0.0.6